+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - An option to decide whether or not to use `sudo` when calling the sensor script
 - Spanish Translation (thanks to [rafagale](https://github.com/rafagale))
+- The ability to set a temperature and humidity offset
 
 ### Fixed
 - The DOM may try to load when there is no sensor data